better memory footprint

+++ image/image.h
@@ -12,8 +12,8 @@
 
 typedef int32_t ColorVal;  // used in computations
 
-typedef int16_t ColorVal_intern_8;   // used in representations
-typedef int32_t ColorVal_intern_16;  // making them signed and a large enough for big palettes and negative alpha values for frame combination
+typedef uint8_t ColorVal_intern_8;
+typedef uint16_t ColorVal_intern_16;
 typedef int32_t ColorVal_intern_32;
 
 // It's a part of C++14. Following impl was taken from GotW#102
